+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # Changelog
 
+## v0.0.2 rpetit3/camlhmp "" 2024/??/??
+
+### `Added`
+
+- `camlhmp-extract` command to extract targets from a set of reference sequences
+- `excludes` field to `profiles` section to allow for exclusion of targets from a profile
+
 ## v0.0.1 rpetit3/camlhmp "Not even walking yet" 2024/04/24
 
 This is a development release for getting things on PyPi and Bioconda. Not expected to be stable.